+++ extensions/slack/src/shared/client/SlackClient.ts
@@ -10,6 +10,11 @@ interface Item {
 export interface User extends Item {
   username: string;
   conversationId: string | undefined;
+  title: string;
+  statusEmoji: string | undefined;
+  statusText: string | undefined;
+  statusExpiration: string;
+  timezone: string;
 }
 
 export type Channel = Item;
@@ -53,10 +58,24 @@ export class SlackClient {
         .filter(
           ({ is_bot, is_workflow_bot, deleted, id }) => !is_bot && !is_workflow_bot && !deleted && id !== "USLACKBOT",
         )
-        .map(({ id, name: username, profile, team_id }) => {
+        .map(({ id, name: username, profile, team_id, tz }) => {
           const firstName = profile?.first_name ?? "";
           const lastName = profile?.last_name ?? "";
           const name = `${firstName} ${lastName}`;
+          const jobTitle = profile?.title ?? "";
+          const statusEmoji = profile?.status_emoji ?? undefined;
+          const statusText = profile?.status_text?.replace(/&amp;/g, "&") ?? undefined;
+          const statusExpiration = profile?.status_expiration ?? undefined;
+          const timezone = tz ?? "";
+          
+          let statusExpirationDate = "";
+          if (statusExpiration) {
+            const date = new Date(statusExpiration * 1000);
+            if (!isNaN(date.getTime())) {
+              statusExpirationDate = date.toISOString().split("T")[0];
+              statusExpirationDate = "Back on " + statusExpirationDate;
+            }
+          }
 
           const displayName = [name, profile?.display_name, profile?.real_name].find((x): x is string => !!x?.trim());
 
@@ -68,7 +87,12 @@ export class SlackClient {
             icon: profile?.image_24,
             teamId: team_id,
             username,
+            title: jobTitle,
+            statusEmoji,
+            statusText,
+            statusExpiration: statusExpirationDate,
             conversationId: conversation?.id,
+            timezone
           };
         })
         .filter((i): i is User => !!(i.id?.trim() && i.name?.trim() && i.teamId?.trim()))